+++ CHANGELOG.md
@@ -50,6 +50,9 @@ For older change log history see the [historic changelog](HISTORIC_CHANGELOG.md)
 
 ### Changed
 
+- SqlServerEndpointPermission
+  - The parameter `ServerName` is now non-mandatory and defaults to `$env:COMPUTERNAME`
+    ([issue #319](https://github.com/dsccommunity/SqlServerDsc/issues/319)).
 - SqlServerLogin
   - The parameter `ServerName` is now non-mandatory and defaults to `$env:COMPUTERNAME`
     ([issue #319](https://github.com/dsccommunity/SqlServerDsc/issues/319)).